+++ packages/expo-manifests/android/src/main/java/expo/modules/manifests/core/ExpoUpdatesManifest.kt
@@ -33,23 +33,8 @@ class ExpoUpdatesManifest(json: JSONObject) : Manifest(json) {
   @Throws(JSONException::class)
   override fun getBundleURL(): String = getLaunchAsset().require("url")
 
-  @Deprecated(message = "exposdk:... runtime version is deprecated")
-  private fun getSDKVersionFromRuntimeVersion(): String? {
-    val runtimeVersion = getRuntimeVersion()
-    if (runtimeVersion == "exposdk:UNVERSIONED") {
-      return "UNVERSIONED"
-    }
-
-    val expoSDKRuntimeVersionRegex: Pattern = Pattern.compile("^exposdk:(\\d+\\.\\d+\\.\\d+)$")
-    val expoSDKRuntimeVersionMatch: Matcher = expoSDKRuntimeVersionRegex.matcher(runtimeVersion)
-    if (expoSDKRuntimeVersionMatch.find()) {
-      return expoSDKRuntimeVersionMatch.group(1)!!
-    }
-    return null
-  }
-
   override fun getExpoGoSDKVersion(): String? {
-    return getExpoClientConfigRootObject()?.getString("sdkVersion") ?: getSDKVersionFromRuntimeVersion()
+    return getExpoClientConfigRootObject()?.getString("sdkVersion")
   }
 
   @Throws(JSONException::class)