+++ packages/expo-manifests/android/src/test/java/expo/modules/manifests/core/ExpoUpdatesManifestTest.kt
@@ -8,9 +8,8 @@ class ExpoUpdatesManifestTest {
   @Test
   @Throws(Exception::class)
   fun testGetSDKVersionNullable_ValidCases() {
-    val runtimeVersion = "exposdk:39.0.0"
     val manifestJson =
-      "{\"runtimeVersion\":\"$runtimeVersion\"}"
+      "{\"extra\":{\"expoClient\":{\"sdkVersion\":\"39.0.0\"}}}"
     val manifest = ExpoUpdatesManifest(JSONObject(manifestJson))
     Assert.assertEquals(manifest.getExpoGoSDKVersion(), "39.0.0")
   }
@@ -18,29 +17,9 @@ class ExpoUpdatesManifestTest {
   @Test
   @Throws(Exception::class)
   fun testGetSDKVersionNullable_ValidCaseUnversioned() {
-    val runtimeVersion = "exposdk:UNVERSIONED"
     val manifestJson =
-      "{\"runtimeVersion\":\"$runtimeVersion\"}"
+      "{\"extra\":{\"expoClient\":{\"sdkVersion\":\"UNVERSIONED\"}}}"
     val manifest = ExpoUpdatesManifest(JSONObject(manifestJson))
     Assert.assertEquals(manifest.getExpoGoSDKVersion(), "UNVERSIONED")
-  }
-
-  @Test
-  @Throws(Exception::class)
-  fun testGetSDKVersionNullable_NotSDKRuntimeVersionCases() {
-    val runtimeVersions = listOf(
-      "exposdk:123",
-      "exposdkd:39.0.0",
-      "exposdk:hello",
-      "bexposdk:39.0.0",
-      "exposdk:39.0.0-beta.0",
-      "exposdk:39.0.0-alpha.256"
-    )
-    runtimeVersions.forEach { runtimeVersion ->
-      val manifestJson =
-        "{\"runtimeVersion\":\"$runtimeVersion\"}"
-      val manifest = ExpoUpdatesManifest(JSONObject(manifestJson))
-      Assert.assertNull(manifest.getExpoGoSDKVersion())
-    }
   }
 }